Implemented GameController with fully functional CRUD endpoints:
GET /Games, POST /Games, DELETE /Games/{id} and PUT /Games/{id}

+++ src/main/java/com/gamecatalog/game_catalog_api/model/Game.java
@@ -49,7 +49,7 @@ public class Game {
         this.developer = developer;
     }
 
-    public long getReleaseYear() {
+    public int getReleaseYear() {
         return releaseYear;
     }
